Conflicts:
	csAnt-setup.sh
	csAnt.node
	scripts/CyclePublish.cs
	scripts/Publish.cs
	scripts/RunTests.cs
	scripts/Test_HelloWorld.cs
	src/SoftwareMonkeys.csAnt.Contracts/SoftwareMonkeys.csAnt.Contracts.node
	src/SoftwareMonkeys.csAnt.IO/FileFinder.cs
	src/SoftwareMonkeys.csAnt.IO/FilesGrabber.cs
	src/SoftwareMonkeys.csAnt.IO/SoftwareMonkeys.csAnt.IO.node
	src/SoftwareMonkeys.csAnt.Projects.Tests.Scripting/SoftwareMonkeys.csAnt.Projects.Tests.Scripting.node
	src/SoftwareMonkeys.csAnt.Projects.Tests/Integration/NodesTestFixture.cs
	src/SoftwareMonkeys.csAnt.Projects.Tests/SoftwareMonkeys.csAnt.Projects.Tests.csproj
	src/SoftwareMonkeys.csAnt.Projects.Tests/SoftwareMonkeys.csAnt.Projects.Tests.node
	src/SoftwareMonkeys.csAnt.Projects/SoftwareMonkeys.csAnt.Projects.node
	src/SoftwareMonkeys.csAnt.SetUpFromLocalConsole/SoftwareMonkeys.csAnt.SetUpFromLocalConsole.node
	src/SoftwareMonkeys.csAnt.SetUpFromWebConsole/Main.cs
	src/SoftwareMonkeys.csAnt.SetUpFromWebConsole/SoftwareMonkeys.csAnt.SetUpFromWebConsole.csproj
	src/SoftwareMonkeys.csAnt.Tests.Scripting/SoftwareMonkeys.csAnt.Tests.Scripting.node
	src/SoftwareMonkeys.csAnt.Tests/NUnitTestRunner.cs
	src/SoftwareMonkeys.csAnt.Tests/SoftwareMonkeys.csAnt.Tests.node
	src/SoftwareMonkeys.csAnt.Tests/Unit/FileFinderTestFixture.cs
	src/SoftwareMonkeys.csAnt.Versions/SoftwareMonkeys.csAnt.Versions.node
	src/SoftwareMonkeys.csAnt/SoftwareMonkeys.csAnt.node
	src/Source.node

+++ scripts/CopyBinToRoot.cs
@@ -1,3 +1,5 @@
+//css_ref ../lib/csAnt/bin/Release/net-40/SoftwareMonkeys.csAnt.Projects.dll
+
 using System;
 using System.IO;
 using System.Diagnostics;
@@ -17,22 +19,29 @@ class CopyBinToRootScript : BaseProjectScript
 			+ Path.DirectorySeparatorChar
 			+ "bin";
 	
-	        Console.WriteLine("");
-	        Console.WriteLine("Bin directory:");
-	        Console.WriteLine(binDirectory);
-	        Console.WriteLine("");
+        Console.WriteLine("");
+        Console.WriteLine("Bin directory:");
+        Console.WriteLine(binDirectory);
+        Console.WriteLine("");
 
 		int i = 0;
+        int s = 0;
+
+        var arguments = new Arguments(args);
+
+        var buildMode = "Release";
+        if (arguments.Contains("mode"))
+            buildMode = arguments["mode"];
+
+        Console.WriteLine("Build mode: " + buildMode);
 
 		var patterns = new string[]{
-			"bin/Release/packed/csAnt-SetUpFromLocal.exe",
-			"bin/Release/packed/csAnt-SetUp.exe"
+			"bin/" + buildMode + "/packed/csAnt-SetUpFromLocal.*",
+			"bin/" + buildMode + "/packed/csAnt-SetUp.*"
 		};
 
 	        foreach (string file in FindFiles(patterns))
 	        {
-		        i++;
-
 		        string toFile = CurrentDirectory
 			        + Path.DirectorySeparatorChar
 			        + Path.GetFileName(file);
@@ -48,10 +57,20 @@ class CopyBinToRootScript : BaseProjectScript
 			        + toFile.Replace(ProjectDirectory, "")
 		        );
 
-		        File.Copy(file, toFile, true);
+                var isNewer = File.GetLastWriteTime(file) > File.GetLastWriteTime(toFile);
+
+                if (!File.Exists(toFile) || isNewer)
+                {
+    		        File.Copy(file, toFile, true);
+
+    		        i++;
+                }
+                else
+                    s++;
 	        }
 
 		Console.WriteLine(i + " files copied.");
+		Console.WriteLine(s + " files skipped.");
 
 		AddSummary("Moved " + i + " files from '/bin/*' to '/'");
